Background_Removal
The aim of this project is to develop a Python-based solution for removing the background from scanned facades dating from 1924 to 1940 to improve their readability and digital preservation. Many historical documents are scanned in environments where uneven lighting or noise can affect the clarity of the print or the paper upon which they have been drawn is showing signs of decay and mold. By employing digital image processing techniques such as thresholding, edge detection, and morphological operations, we will isolate the textual content from the background. This will help enhance the legibility of faded lines, remove artifacts, and prepare documents for further deep learning applications. The project will utilize libraries like OpenCV and PIL to perform image preprocessing, followed by background subtraction algorithms to create cleaner, more readable document images. Ultimately, this approach aims to automate the restoration of archived architectural drawings, making them more accessible and usable for digital archival purposes. The dataset employed for this project is formatted in 1024x1024 pixel format and features 1784 images of real buildings built in the city of Thessaloniki, Greece designed in the mid-war era.